RETRY:
Added features to the profile page. A div that contains the discriptions of the location and a seperate div for specific events happening at said location.

Also added some JS functionality so the div mentioned above are pushed to the side when the navbar opens.

Added CSS accordingly, and all header and title tags now dafault to "Big John"

Fixed thymeleaf syntax, so everything(partials/templates work as they should.